requirements.txt
streamlit
sqlalchemy
psycopg2
requests
beautifulsoup4
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
numpy